refactor(Server): rename createFeatures to setupFeatures
