+++ lib/Server.js
@@ -31,7 +31,7 @@ const webpackDevMiddleware = require('webpack-dev-middleware');
 const updateCompiler = require('./utils/updateCompiler');
 const createLogger = require('./utils/createLogger');
 const createCertificate = require('./utils/createCertificate');
-const createFeatures = require('./utils/createFeatures');
+const setupFeatures = require('./utils/setupFeatures');
 const routes = require('./utils/routes');
 
 const validateOptions = require('schema-utils');
@@ -207,7 +207,7 @@ class Server {
     // Keep track of websocket proxies for external websocket upgrade.
     const websocketProxies = [];
 
-    const features = createFeatures.call(this, {
+    const features = setupFeatures.call(this, {
       app,
       compiler,
       options,
