+++ lib/utils/setupFeatures.js
